requirements.txt
transformers==4.36.0
datasets==2.14.0
torch==2.1.0
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
beautifulsoup4==4.12.2
requests==2.31.0
rouge-score==0.1.2
matplotlib==3.7.2
seaborn==0.12.2
wandb==0.15.12
newspaper3k==0.2.8
tqdm==4.66.1
scikit-learn==1.3.0
sentencepiece==0.1.99
lxml>=4.9.0
pillow==10.0.0
nltk==3.8.1